requirements.txt
Django==1.6.4
Markdown==2.4
django-filter==0.7
djangorestframework==2.3.13
drf-nested-routers==0.1.3
jsonfield==0.9.20
wsgiref==0.1.2
